Add files via upload
You are given link of IPL 2020-2021 home page link. https://www.espncricinfo.com/series/ipl-2020-21-1210595. You need to scrap data using this link in the format displayed below.

Create a folder of ipl 2020.
Inside that folder create folders for each team playing in that ipl.
Create Excel file of each batsman in that team folder in which we will store details of the player in the form of array of object where each object should describe a particular match played by a player. Details of an object includes a. Runs, balls, sixes , fours, strike rate (sr) for that match b. date ,venue ,result and opponent name for that match